+++ app/admin/ecommerce/products/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { fetcher, useFetcher } from "../../../components/global/fetcher";
 import Loading from "../../../components/global/loading";
 import { useAtom } from "jotai";
@@ -11,8 +11,10 @@ import LightDataGrid from "@/app/components/global/LightDataGrid/LightDataGrid";
 import { Button, IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 import ModeEditIcon from "@mui/icons-material/ModeEdit";
+import Swal from "sweetalert2";
 export default function Products() {
   const [title, setTitle] = useAtom(pageTitle);
+  const [triggered, setTriggered] = useState(false);
 
   useEffect(() => {
     setTitle({
@@ -24,12 +26,25 @@ export default function Products() {
 
   const deleteRow = async (id) => {
     try {
-      const req = await fetcher({
-        url: `/v1/api/ecommerce/admin/products/${id}`,
-        method: "DELETE",
+      const result = await Swal.fire({
+        title: "مطمئن هستید؟",
+        text: "با حذف این گزینه امکان بازگشت آن وجود ندارد",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "بله حذفش کن",
+        cancelButtonText: "لغو",
       });
-      toast.success("موفق");
-      refetchProducts();
+
+      if (result.isConfirmed) {
+        const req = await fetcher({
+          url: `/v1/api/ecommerce/admin/products/${id}`,
+          method: "DELETE",
+        });
+        toast.success("موفق");
+        setTriggered(!triggered);
+      }
     } catch (error) {
       toast.error(error.message);
     }
@@ -43,32 +58,57 @@ export default function Products() {
       maxSize: 10,
     },
     {
+      accessorKey: "pPhoto",
+      header: "تصویر",
+      minSize: 100,
+      maxSize: 100,
+      size: 100,
+      Cell: ({ row }) => {
+        return row?.original?.attachments?.length ? (
+          <Image
+            key={row.id}
+            loading="eager"
+            src={`${
+              process.env.NEXT_PUBLIC_BASE_URL
+            }/v1/api/ecommerce/productphotos/image/${
+              row.original.attachments[0]?.fileName || ""
+            }`}
+            width={30}
+            height={30}
+            alt=""
+          />
+        ) : (
+          <img width={30} height={30} src="/images/no-photo.png" alt="" />
+        );
+      },
+    },
+    {
       accessorKey: "title",
       header: "نام ",
-      minSize: 100, //min size enforced during resizing
-      maxSize: 400, //max size enforced during resizing
-      size: 400, //medium column
+      minSize: 100,
+      maxSize: 400,
+      size: 400,
     },
     {
       accessorKey: "brand.name",
       header: "برند",
-      minSize: 100, //min size enforced during resizing
-      maxSize: 150, //max size enforced during resizing
-      size: 150, //medium column
+      minSize: 100,
+      maxSize: 150,
+      size: 150,
     },
     {
       accessorKey: "publishStatus.name",
       header: "وضعیت انتشار",
-      minSize: 100, //min size enforced during resizing
-      maxSize: 150, //max size enforced during resizing
-      size: 150, //medium column
+      minSize: 100,
+      maxSize: 150,
+      size: 150,
     },
     {
       accessorKey: "entityType.name",
       header: "دسته بندی",
-      minSize: 100, //min size enforced during resizing
-      maxSize: 400, //max size enforced during resizing
-      size: 180, //medium column
+      minSize: 100,
+      maxSize: 400,
+      size: 180,
     },
 
     {
@@ -81,12 +121,7 @@ export default function Products() {
       accessorKey: "Actions",
       header: "عملیات",
       size: 200,
-      muiTableHeadCellProps: {
-        align: "right",
-      },
-      muiTableBodyCellProps: {
-        align: "right",
-      },
+
       Cell: ({ row }) => (
         <>
           <a href={`/admin/ecommerce/products/${row.id}`}>
@@ -109,6 +144,7 @@ export default function Products() {
       <LightDataGrid
         url={"/v1/api/ecommerce/admin/products"}
         columns={columns}
+        triggered={triggered}
       />
     </div>
   );
